backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
blinker==1.9.0
    # via flask
cachelib==0.13.0
    # via flask-session
click==8.3.1
    # via flask
flask==3.1.3
    # via
    #   -r requirements.in
    #   flask-session
    #   flask-sqlalchemy
flask-session==0.8.0
    # via -r requirements.in
flask-sqlalchemy==3.1.1
    # via -r requirements.in
greenlet==3.3.2
    # via sqlalchemy
iniconfig==2.3.0
    # via pytest
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
msgspec==0.20.0
    # via flask-session
packaging==26.0
    # via pytest
pluggy==1.6.0
    # via pytest
pydantic==2.12.5
    # via -r requirements.in
pydantic-core==2.41.5
    # via pydantic
pygments==2.20.0
    # via pytest
pytest==9.0.2
    # via -r requirements.in
python-dotenv==1.2.2
    # via -r requirements.in
sqlalchemy==2.0.48
    # via flask-sqlalchemy
typing-extensions==4.15.0
    # via
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
werkzeug==3.1.7
    # via flask
